requirements.txt
aiosqlite==0.22.1
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
datasketch==1.9.0
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.23
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
Levenshtein==0.27.3
numpy==2.4.3
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-frontmatter==1.1.0
python-Levenshtein==0.27.3
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.33.0
requests-toolbelt==1.0.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
sqlite-vec==0.1.9
tenacity==9.1.4
thefuzz==0.22.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
